feature: change to hpp and add cmake script
